pyproject.toml
[tool.poetry]
name = "mark4"
version = "0.1.0"
description = ""
authors = ["Jon-Cody Sokoll <sokoll.jc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
tensorflow = "^2.12.0"
hydra-core = "^1.3.2"
tensorflow-datasets = "^4.8.3"
pytest = "^7.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
